Updates to deal with gcc warnings when compiling in Debug
git-svn-id: svn+ssh://cyverse.warwick.ac.uk/home/svn/TRegS_software/LiBiNorm/trunk@2280 d27ce7f2-3f32-4a89-82b7-fbde2a400052

+++ LiBiNormSrc/LiBiCount.h
@@ -21,7 +21,7 @@ class LiBiCount
 	size_t bamCounter,maxCacheSize;
 	int minqual;
 
-	void incBamCounter(const BamAlignment * ba = 0,size_t size = -1);
+	void incBamCounter(const BamAlignment * ba = 0,int size = -1);
 
 public:
 	bool useStrand,reverseStrand,verbose,htSeqCompatible,nameOrder;
